requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.1.4
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
Faker==1.0.0
gunicorn==19.9.0
idna==2.7
PyJWT==1.7.0
python-dateutil==2.7.5
python-dotenv==0.10.0
pytz==2018.7
requests==2.20.1
six==1.11.0
text-unidecode==1.2
urllib3==1.24.1
